+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to insta and cargo-insta are documented here.
 ## 1.7.0
 
 * Added support for u128/i128.  (#169)
+* Normalize newlines to unix before before asserting.  (#172)
+* Switch diffing to patience.  (#173)
 
 ## 1.6.3